paddle2tlx-vision/requirements.txt
astor==0.8.1
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
h5py==3.7.0
imageio==2.19.2
numpy==1.21.6
opencv-python==4.5.5.64
pandas==1.3.5
Pillow==9.1.1
protobuf==3.20.1
requests==2.27.1
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
urllib3==1.26.9

#Keras-Applications==1.0.8
#Keras-Preprocessing==1.1.2
#tensorflow==2.0.0
#torch==1.6.0  # pip install torch==1.6.0+cpu -f https://download.pytorch.org/whl/torch_stable.html
#torch==1.6.0+cu101
#torchvision==0.7.0+cu101
#torchvision==0.7.0+cpu
#torch==1.10.0+cpu
#torchvision==0.11.1+cpu
#torchsummary==1.5.1
paddlepaddle==2.3.0  # 2.2.0
tensorlayerx==0.5.7  # 0.5.3

#onnx==1.11.0
#onnxruntime==1.11.1
#x2paddle==1.2.0
#tlx2onnx==0.0.1a0